requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.2.7
astroid==2.4.2
async-lru==1.0.2
atomicwrites==1.4.0
attrs==19.3.0
bottle==0.12.18
bottle-websocket==0.2.9
certifi==2020.4.5.1
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coverage==5.3
cryptography==3.1
curl==0.0.1
decorator==4.4.2
distlib==0.3.1
Django==3.0.5
django-cors-headers==3.5.0
django-debug-toolbar==3.1.1
django-filter==2.4.0
django-outbox==0.6.0
djangorestframework==3.11.0
djangorestframework-filters==0.11.1
djangorestframework-simplejwt==4.6.0
Eel==0.14.0
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
future==0.18.2
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==0.4.17
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
iniconfig==1.0.1
isort==5.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
mypy==0.790
mypy-extensions==0.4.3
packaging==20.3
Pillow==8.0.1
pluggy==0.13.1
py==1.8.1
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
Pyrogram==1.0.7
PySocks==1.7.1
pyTelegramBotAPI==3.7.3
pytest==5.4.1
pytest-django==3.9.0
python-dotenv==0.13.0
python-telegram-bot==12.7
pytils==0.3
pytz==2019.3
requests==2.23.0
service-identity==18.1.0
six==1.14.0
sorl-thumbnail==12.6.3
SQLAlchemy==1.3.20
sqlparse==0.3.1
TelegramBotAPI==0.3.2
TgCrypto==1.2.1
toml==0.10.1
tornado==6.0.4
twilio==6.48.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.9
virtualenv==20.0.31
wcwidth==0.1.9
Werkzeug==1.0.1
whichcraft==0.6.1
wrapt==1.12.1
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.2.0
